Add read ahead for VolumeReaderJavaScriptStream.
Read ahead archive's data from JavaScript in order to not block for it
when libarchive requests it. This improvement is usefull only for
consecutive reads. In case of Seek / Skip the read ahead data will be
discarded.

Performance improvement is around 7-10% on Pixel for a compressed file of
232 MB to 30 MB inside a zip archive.

BUG=chromium:388077

+++ unpacker/cpp/javascript_requestor.h
@@ -16,7 +16,7 @@ class JavaScriptRequestor {
   // Request a file chunk from JavaScript. The request is asynchronous.
   virtual void RequestFileChunk(const std::string& request_id,
                                 int64_t offset,
-                                int32_t bytes_to_read) = 0;
+                                size_t bytes_to_read) = 0;
 };
 
 #endif  // JAVASCRIPT_REQUESTOR_H_
