+++ unpacker/cpp/request.cc
@@ -35,14 +35,17 @@ pp::VarDictionary request::CreateReadChunkRequest(
     const std::string& file_system_id,
     const std::string& request_id,
     int64_t offset,
-    int32_t length) {
+    size_t length) {
   pp::VarDictionary request =
       CreateBasicRequest(READ_CHUNK, file_system_id, request_id);
 
   std::stringstream ss_offset;
   ss_offset << offset;
   request.Set(request::key::kOffset, ss_offset.str());
-  request.Set(request::key::kLength, length);
+
+  // TODO(cmihail): Think of using either size_t, int32_t or int64_t everywhere.
+  // The static cast is just a temporary solution to avoid big CL.
+  request.Set(request::key::kLength, static_cast<int32_t>(length));
   return request;
 }
